grammatical/punctuation edits, completed sentence
At line 49, the paragraph ended with a sentence that cut itself off: "Preferable are standards that have been " (sic). I changed this to the sentence: "It is preferable to use standards that signify heading levels, as these standards are not only machine-readable, but also translate easily across web browsers and potential future content migrations." I am not sure whether it is prudent to mention HTML or CSS standards by name, but I think I approximated the intent of the original sentence as best as I was able. I also added some necessary commas where they were missing from a number of compound-complex sentences, in order to improve readability.